Resolve bug that caused resolution errors when multiple references to the same mob existed in a file. The ResolveRefVisitor.cpp::GetChildComponents function was considering all references when it should have been considering only contained objects.
Restore regression test to workign order. Resolved several problems in
recent code changes exposed by reactivating the tests. Update
EPTrackContentsTest to include MobSlot::UserPos test. Improved layout
of TestFile/GNUmakefile by factoring out repeated boiler plate test
code.

Updated the requirements xml element to include "annotation" and
"note" sub elements and an "action" attribute. The action attribute is
a place to record the fact that something needs to be done to bring
the requirement in the xml file into sync with the written spec. The
annotation is intended to record some additional text that is proposed
(or agreed) should go into the spec. The note is there to comment on
the implementation or interpretation of the spec.

Update the TestResult to include concept of covered and noted test
results. Requirements covered by an associated test are now
initialized as covered requirements.  From there they move to noted,
passed, warned, or failed. when the detailed test results are
consolidated. This is a step towared reporting requirements that are
covered by a particular file rather than simple reporting all
requirements associated with a test as passed (which wasn't the
original goal but was the original implementation).

Resolve tracker item 1388284 "Missclassifed Requirements". Change
REQ_EP_013,014,039 to app requirements.

Resolve tracker item 1388314 "Test MIDI files". Changed REQ_EP_143 to
app requirment because the analyzer is not responsible for verifying
the contents of the external midi file. And, partially resolve 1388290
- REQ_EP_023 should be app requirement not file. And partially resolve
tracker item 1388290 - REQ_EP_097 - Pulldown may be used... should be
app requirement. And resolve tracker item 1388299 - REQ_EP_160
misclassifed, changed to app.

Resolve tracker item 1388140 ] "TimelineMobSlot Edit Points". Split
REQ_EP_108 into .1/.2/.3/.4 parts to isolate individual reporting
items. Updated the test to note the presence of mark in/out points and
user position. The only thing tested is that mark out is greater than
mark in if both exist. No long warn if one of mark in/out is present
but not the other.