Split read-portion of this test into two stages: 1. Reading records back using only GetValue() style calls, no struct offsets registered. 2. Reading records back using a mix of GetValue() and GetStruct() with struct offsets registered.
This is necessary to obtain a complete read-test when working with files
that are not Structured Storage encoding or when working in eager SS loading
mode.  In these other modes, default offsets are used by the TypeDefRecord
during file loading which precludes registration of user-defined offsets
(RegisterOffsets returns DEFAULT_ALREADY_USED). Returns of
DEFAULT_ALREADY_USED causes the rest of the test to be skipped and the test
is regarded as successful.

Each of these two styles of read tests are now performed in both eager and
lazy loading modes.